This is commit for efforts to convert dyGiLa-paraIO-Tfield branch to GPU workable code. The changes have been done are (1) matep.hpp member functions are marked #pragma hila loop_function     This preprocessing directive marks memeber functions as loop_fucntion,     such as hilapp could convert marked functions to __host__ __device__; (2) This modified matep then been used in glsol::next* and glsol::write* funtions.     It is used in onsites block for Temerature field;
Currently, the code can't be compiled. The design of matep and maybe also glsol
aren't compatible with HILA's methodlogy of GPU computing, many functions in matep
induce new errors as device codes of HILA's type. And the data arrays matep hold are
also incompatible with HILA. Here is comment and suggestion from Kari:

there may be more problems which prevent straightforward gpu conversion. The functions called from
onsite() -loops may not do i/o or use global or static state, i.e. they should be (almost) pure functions.  Thus, if the matep functions use arrays to store the state, these should be passed as arguments to functions.  This is due to the fact that these must be copied to the gpu device, and hilapp does this only if the variables are visible on the top level, i.e. in
onsite(). I have had plans to guide hilapp to do deeper inspection of loop functions, but this rapidly becomes complicated.
i/o from within onsites() is not supported at all. The data should be copied out by other means (which then depends on the kind of data needed).